considered window already opened
