+++ apps/remix-ide/src/app/components/popup-panel.tsx
@@ -4,6 +4,7 @@ import { PluginRecord, RemixPluginPanel } from '@remix-ui/panel'
 import packageJson from '../../../../../package.json'
 import { PluginViewWrapper } from '@remix-ui/helper'
 import { EventEmitter } from 'events'
+import DOMPurify from 'dompurify';
 
 import { AppAction, appActionTypes, AppState } from '@remix-ui/app'
 
@@ -23,6 +24,7 @@ export class PopupPanel extends AbstractPanel {
   element: HTMLDivElement
   dispatch: React.Dispatch<any> = () => { }
   appStateDispatch: React.Dispatch<AppAction> = () => { }
+  hooks: boolean = false
 
   constructor(config = null) {
     super(profile)
@@ -64,7 +66,6 @@ export class PopupPanel extends AbstractPanel {
   }
 
   async showPopupPanel(show) {
-
     this.appStateDispatch({
       type: appActionTypes.setShowPopupPanel,
       payload: show
@@ -85,6 +86,34 @@ export class PopupPanel extends AbstractPanel {
   }
 
   updateComponent(state: popupPanelState, appState: Partial<AppState>) {
+    if (!this.hooks) {
+      try {
+        const markdown = document.getElementsByClassName('nlux-composer-container')
+        const button = markdown[0].getElementsByTagName('button')[0]
+        const textArea = markdown[0].getElementsByTagName('textarea')[0]
+        // only add event listeners if they are not already added
+        if (!textArea.dataset.listenerAdded) {
+          textArea.addEventListener('input', (event) => {
+            const sanitizedInput = DOMPurify.sanitize(textArea.value)
+            if (sanitizedInput !== textArea.value) {
+              textArea.value = sanitizedInput
+            }
+          })
+          textArea.dataset.listenerAdded = 'true'
+        }
+
+        if (!button.dataset.listenerAdded) {
+          button.dataset.listenerAdded = 'true'
+          button.addEventListener('click', (event) => {
+            const sanitizedInput = DOMPurify.sanitize(textArea.value)
+            if (sanitizedInput !== textArea.value) {
+              textArea.value = sanitizedInput
+            }
+          })
+        }
+        this.hooks = true
+      } catch (error) { this.hooks = false }
+    }
     return (
       <div
         className={`px-0 bg-light border-info ${appState?.showPopupPanel ? 'd-flex' : 'd-none'}`}
